Exercise 3 and 4 working
Rounded up Exercise 3 which is run by escript exercise 3. Made some
changes to myAffine so the parallelograms are depicted corrected and the
input is more convenient. Take care of which points to assign where. x
and y are in the rotated image. xx and yy should be in  the original
one. (this is still somewhat mixed up in the function. Can be run from
runAffine.